Progress search by Town
 - had to add a csv upload because existing one failed to read response

+++ src/Tests/sfa.Tl.Marketing.Communication.Tests/Builders/TownEntityListBuilder.cs
@@ -22,8 +22,8 @@ public class TownEntityListBuilder
                 Name = $"Test Town {nextId}",
                 County = $"County {nextId}",
                 LocalAuthority = $"Local Authority {nextId}",
-                Latitude = decimal.Parse($"50.0{nextId}"),
-                Longitude = decimal.Parse($"-1.0{nextId}"),
+                Latitude = double.Parse($"50.0{nextId}"),
+                Longitude = double.Parse($"-1.0{nextId}"),
                 // ReSharper disable once StringLiteralTypo
                 SearchString = $"testtown{nextId}county{nextId}",
                 PartitionKey = "T"
